+++ src/engine/Operation.h
@@ -88,7 +88,7 @@ class Operation {
   virtual void setTextLimit(size_t limit) = 0;
   virtual size_t getCostEstimate() = 0;
 
-  virtual size_t getSizeEstimate() final {
+  virtual uint64_t getSizeEstimate() final {
     if (_limit._limit.has_value()) {
       return std::min(_limit._limit.value(), getSizeEstimateBeforeLimit());
     } else {
@@ -97,7 +97,7 @@ class Operation {
   }
 
  private:
-  virtual size_t getSizeEstimateBeforeLimit() = 0;
+  virtual uint64_t getSizeEstimateBeforeLimit() = 0;
 
  public:
   virtual float getMultiplicity(size_t col) = 0;
